FAC-225 Add userId into KitDsl creation and modify database schema
The KitDsl creation has been modified to include the userId, which is now passed as an argument in the KitDslMapper and KitDslPersistenceJpaAdapter. A Liquibase changelog was added to include userId to the KitDsl database schema, apply the name conventions for tables and sequences, and modify some column names and types. Tests are updated accordingly in response to these changes.