FAC-320 Change GetKitQuestionnaireDetailService output
Make GetKitQuestionnaireDetailService to use Result record from GetKitQuestionnaireDetailUseCase. Use the Question domain model in the Result record of both GetKitQuestionnaireDetailUseCase and LoadKitQuestionnaireDetailPort. Define a Question record in GetKitQuestionnaireDetailResponseDto to extract needed fields from the Question domain model which has been returned as the result of GetKitQuestionnaireDetailService in the controller.
Fix the tests in GetKitQuestionnaireDetailServiceTest due to these changes.