FAC-359 Add checks to the service
Check spacedIds state, if its empty or null, then should load the current user spaceIds. If it's not null or empty, then check its size is one, and check current user has access to the space.These checks have implemented,because of the usage of this service in two parts of the application.